pyproject.toml
[project]
name = "repl-nix-webharvestpro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.4",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.40.1",
    "trafilatura>=1.12.2",
]
